Meta: Silence error output from `du`'s `--help` invocation
The BSD-flavored `du`s don't have a --help argument, meaning they fail
the check, which is what we want. But the output from that invocation
goes to stderr which, on the build script, means going to the user's
screen. This can lead people to believe there's an error when that
isn't the case.